+++ python/CHANGELOG.md
@@ -12,6 +12,7 @@ _At the moment, the project does **not** adhere to [Semantic Versioning](https:/
 ### Added
 
 - `trezorctl set unsafe-prompts` controls the new "unsafe prompts" feature.  [#1126]
+- `trezorctl btc get-address` can create multisig addresses.
 
 ### Changed